+++ example/example/lib/stories.dart
@@ -1,0 +1,42 @@
+import 'package:example/stories/zwap_buttons.dart';
+import 'package:example/stories/zwap_percent.dart';
+import 'package:example/stories/zwap_weekly_calendar_pickert.dart';
+import 'package:flutter/material.dart';
+
+import 'package:flutter/material.dart';
+
+enum ZwapStories {
+  buttons,
+  percent,
+  weeklyCalendarPicker,
+}
+
+class ZwapStory {
+  String name;
+  ZwapStories story;
+  Widget Function(BuildContext) builder;
+
+  ZwapStory({
+    required this.name,
+    required this.builder,
+    required this.story,
+  });
+}
+
+List<ZwapStory> stories = [
+  ZwapStory(
+    name: "ZwapButton",
+    builder: (context) => const ZwapButtonsStory(),
+    story: ZwapStories.buttons,
+  ),
+  ZwapStory(
+    name: "ZwapPercent",
+    builder: (context) => const ZwapPercentStory(),
+    story: ZwapStories.percent,
+  ),
+  ZwapStory(
+    name: "ZwapWeeklyCalendarPicker",
+    builder: (context) => const ZwapCalendarPickerStory(),
+    story: ZwapStories.weeklyCalendarPicker,
+  )
+];